% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtext.R
\name{Question-Text}
\alias{Question-Text}
\alias{qtext}
\alias{qtext<-}
\title{Get and Set Variable Question Text}
\usage{
qtext(x)

qtext(x) <- value
}
\arguments{
\item{x}{a single column in a \code{svy_df} object, or a survey data frame of
class \code{svy_df} to return all variable question text for.}

\item{value}{character: A string containing the full wording of a survey item
to assign to the question text attribute of \code{x}.}
}
\value{
An object identical to \code{x} with the 'qtext' attribute containing
  question text, or for \code{questions}, a two column data frame with the
  column names of the data in the first column and the associated question
  text of each in the second.
}
\description{
Generic method for retreiving and assinging the question text associated with
a column in a \code{svy_df} object associated with a survey question.
}
\examples{
\dontrun{
## Show the current question text for a column
qtext(data$nr1)

## Assign the question text to a variable
qtext(data$nr1) <- 'yolo
}
}

