% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svy-df.R
\name{svy_df}
\alias{is.svy_df}
\alias{print.svy_df}
\alias{svy_df}
\title{Create a survey data frame}
\usage{
svy_df(data, questions = NULL)

is.svy_df(x)

\method{print}{svy_df}(x, ..., n = NULL, width = NULL)
}
\arguments{
\item{data}{a data frame}

\item{questions}{a data frame with a minimum of two columns, where the first
column contains matching names of columns in \code{data} and the second
column contains the associated full question text.}

\item{x}{An object to test or print.}

\item{n,width,...}{Arguemnts passed to \code{print}.}
}
\value{
A object of class \code{svy_df} with optional attributes for question
  text and variable levels.
}
\description{
A survey data frame that employs the same generic methods as a regular
\code{data.frame} with additional functionality for preserving levels and
question text attributes for columns.
}
\details{
Each column of a \code{svy_df} generally represents a question from a opinion
survey, where the values represent unique response choices, variable levels
represent the response choice text, and the question text represents the full
wording of the survey item as displayed to the respondent. Each row
represents a unique respondent.

When subsetting a \code{svy_df} object using standard indexing methods (i.e.
\code{'['}), variable levels and question text will be copied from the
original object to the subsetted object using \code{\link{copy_lev}}.
}

