% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svy-tbl.R
\name{svy_tbl}
\alias{svy_tbl}
\alias{table.svy_df}
\title{Create a table from a survey data frame}
\usage{
\method{table}{svy_df}(data, x, y = NULL, w = NULL, prop = FALSE,
  digits = 0)
}
\arguments{
\item{data}{A survey data frame created by \code{\link{svy_df}}.}

\item{x,y}{character: Column names in \code{data} to tabualte. If \code{y} is
not \code{NULL}, creates a two-way frequency table.}

\item{w}{character: Optional column name containing a vector of survey weights.}

\item{prop}{logical: If \code{TRUE}, print proportions instead of counts.}

\item{digits}{numeric: Number of digits to round counts to.}
}
\value{
A (weighted) frequency table of class \code{svy_tbl}.
}
\description{
Quick and simple methods for creating a weighted frequency table for columns
in a \code{svy_df} object. Printing the table will return formatted output
with the question text of \code{x}.
}
\examples{
\donttest{
## Make sure `data` is a svy_df
is.svy_df(data)

## Make a one-way frequency table
my.tab1 <- table(data, "nr2", w = "wts")

## Make a table of two variables
my.tab2 <- table(data, "nr2", "demGender", "wts")

## Print formatted output
my.tab2
}
}

